requirements.txt
fastapi
uvicorn
groq
edge-tts
python-multipart
python-dotenv
pydantic>=2.0.0
anyio
httpx
distro
tenacity
aiofiles
PyPDF2
